Updated Invoice CRUD Functions
